Renaming vadd_mm to vadd_s as it is a streaming kernel Renaming vadd to vadd_mm as it is a memory mapped kernel
