+++ Developer_Contributed/01-Versal_Custom_Thin_Platform_Extensible_System/ip/s2mm_vadd_s/src/s2mm_vadd_s.cpp
@@ -23,21 +23,19 @@ limitations under the License.
 
 extern "C" {
 
-void mm2s_vadd_mm(ap_int<WORD_SIZE>* mem, hls::stream<hls::axis<ap_int<WORD_SIZE>, 0, 0, 0>>& s, int size) {
+void s2mm_vadd_s(ap_int<WORD_SIZE>* mem, hls::stream<qdma_axis<WORD_SIZE, 0, 0, 0>  >& s, int size) {
 #pragma HLS INTERFACE m_axi port=mem offset=slave bundle=gmem
 
-#pragma HLS INTERFACE axis port=s
+#pragma HLS interface axis port=s
 
 #pragma HLS INTERFACE s_axilite port=mem bundle=control
 #pragma HLS INTERFACE s_axilite port=size bundle=control
-#pragma HLS INTERFACE s_axilite port=return bundle=control
+#pragma HLS interface s_axilite port=return bundle=control
 
 	for(int i = 0; i < size; i++) {
 #pragma HLS PIPELINE II=1
-		hls::axis<ap_int<WORD_SIZE>, 0, 0, 0> x;
-		x.data = mem[i];
-		x.keep = -1;
-		s.write(x);
+		qdma_axis<WORD_SIZE, 0, 0, 0> x = s.read();
+		mem[i] = x.data;
 	}
 
 }
