+++ Developer_Contributed/01-Versal_Custom_Thin_Platform_Extensible_System/ip/vadd_mm/src/vadd_mm.cl
@@ -18,7 +18,7 @@
 
 kernel __attribute__((reqd_work_group_size(1, 1, 1)))
 
-void vadd(global const int* a,
+void vadd_mm(global const int* a,
                    global const int* b,
                    global int* c,
                    const int n_elements)
