+++ Developer_Contributed/01-Versal_Custom_Thin_Platform_Extensible_System/vitis/src/ila_0_bd.cfg
@@ -18,16 +18,16 @@
 
 # System ILA instance
 [debug]
-chipscope=counter_1:m00_axis
-chipscope=counter_1:m01_axis
-chipscope=counter_1:m02_axis
-chipscope=counter_1:m03_axis
-chipscope=counter_1:m04_axis
-chipscope=subtractor_1:s01_axis
-chipscope=subtractor_1:s02_axis
-chipscope=subtractor_1:s03_axis
-chipscope=subtractor_1:s04_axis
-chipscope=subtractor_1:s00_axi
+chipscope=counter_0:m00_axis
+chipscope=counter_0:m01_axis
+chipscope=counter_0:m02_axis
+chipscope=counter_0:m03_axis
+chipscope=counter_0:m04_axis
+chipscope=subtractor_0:s01_axis
+chipscope=subtractor_0:s02_axis
+chipscope=subtractor_0:s03_axis
+chipscope=subtractor_0:s04_axis
+chipscope=subtractor_0:s00_axi
 
 [linkhook]
 custom=postDebugProfile,./../src/ila_0_def.tcl
