+++ Developer_Contributed/01-Versal_Custom_Thin_Platform_Extensible_System/vitis/src/system.cfg
@@ -15,34 +15,34 @@
 #
 
 [connectivity]
-nk=counter:1
-nk=subtractor:1
-nk=vadd:1
+nk=counter:1:counter_0
+nk=subtractor:1:subtractor_0
 nk=vadd_mm:1
-nk=mm2s_vadd_mm:2
-nk=s2mm_vadd_mm:1
+nk=vadd_s:1
+nk=mm2s_vadd_s:2
+nk=s2mm_vadd_s:1
 
-sc=counter_1.m00_axis:subtractor_1.s00_axis
-sc=counter_1.m01_axis:ai_engine_0.DataIn0
-sc=counter_1.m02_axis:ai_engine_0.DataIn1
-sc=counter_1.m03_axis:ai_engine_0.DataIn2
-sc=counter_1.m04_axis:ai_engine_0.DataIn3
-sc=ai_engine_0.DataOut0:subtractor_1.s01_axis
-sc=ai_engine_0.DataOut1:subtractor_1.s02_axis
-sc=ai_engine_0.DataOut2:subtractor_1.s03_axis
-sc=ai_engine_0.DataOut3:subtractor_1.s04_axis
+sc=counter_0.m00_axis:subtractor_0.s00_axis
+sc=counter_0.m01_axis:ai_engine_0.DataIn0
+sc=counter_0.m02_axis:ai_engine_0.DataIn1
+sc=counter_0.m03_axis:ai_engine_0.DataIn2
+sc=counter_0.m04_axis:ai_engine_0.DataIn3
+sc=ai_engine_0.DataOut0:subtractor_0.s01_axis
+sc=ai_engine_0.DataOut1:subtractor_0.s02_axis
+sc=ai_engine_0.DataOut2:subtractor_0.s03_axis
+sc=ai_engine_0.DataOut3:subtractor_0.s04_axis
 
-sc=mm2s_vadd_mm_1.s:vadd_mm_1.in0
-sc=mm2s_vadd_mm_2.s:vadd_mm_1.in1
-sc=vadd_mm_1.out:s2mm_vadd_mm_1.s
+sc=mm2s_vadd_s_1.s:vadd_s_1.in0
+sc=mm2s_vadd_s_2.s:vadd_s_1.in1
+sc=vadd_s_1.out:s2mm_vadd_s_1.s
 
 [clock]
 #id=0  -> 500MHz
-id=0:subtractor_1,counter_1
+id=0:subtractor_0,counter_0
 #id=1  -> 250MHz
-id=1:vadd_1
+id=1:vadd_mm_1
 #id=12 -> 333MHz
-id=12:vadd_mm_1,mm2s_vadd_mm_1,mm2s_vadd_mm_2,s2mm_vadd_mm_1
+id=12:vadd_s_1,mm2s_vadd_s_1,mm2s_vadd_s_2,s2mm_vadd_s_1
 
 [advanced]
 #param=hw_emu.enableProfiling=false
